Use _pop_param_values in es search query modifiers
Previously each modifier had essentially the same code to extact
all values of the same key from the params multidict, now they just
call _pop_param_values(key) instead.

Some of the tests were passing in a dict-type value for params instead of
a MultiDict-type. This change forces the params to now be a MultiDict as it
uses a MultiDict-only method called getall().